v2 endpoints skeleton
