+++ lib/entities/quote-request/IndicativeRequest.ts
@@ -1,101 +1,91 @@
 import { TradeType } from '@uniswap/sdk-core';
-import { BigNumber, ethers } from 'ethers';
-import { getAddress } from 'ethers/lib/utils';
+import { BigNumber, constants, utils } from 'ethers';
+import { V2PostQuoteRequestBody } from '../../handlers/quote-v2';
 
-import { PostQuoteRequestBody } from '../handlers/quote';
-
-export * from './quote-request';
-
-export interface QuoteRequestData {
-  tokenInChainId: number;
-  tokenOutChainId: number;
-  requestId: string;
-  swapper: string;
-  tokenIn: string;
+export interface V2QuoteRequestData extends Omit<V2PostQuoteRequestBody, 'amount' | 'type'> {
   amount: BigNumber;
-  tokenOut: string;
   type: TradeType;
-  numOutputs: number;
   quoteId?: string;
 }
 
-export interface QuoteRequestDataJSON extends Omit<QuoteRequestData, 'amount' | 'type'> {
-  amount: string;
-  type: string;
+export interface V2QuoteRequestJSON extends V2PostQuoteRequestBody {
   quoteId?: string;
 }
 
-// data class for QuoteRequest helpers and conversions
-export class QuoteRequest {
-  public static fromRequestBody(body: PostQuoteRequestBody): QuoteRequest {
-    return new QuoteRequest({
+export class V2QuoteRequest {
+  public static fromRequestBody(body: V2PostQuoteRequestBody): V2QuoteRequest {
+    return new V2QuoteRequest({
       tokenInChainId: body.tokenInChainId,
       tokenOutChainId: body.tokenOutChainId,
       requestId: body.requestId,
-      swapper: getAddress(body.swapper),
-      tokenIn: getAddress(body.tokenIn),
-      tokenOut: getAddress(body.tokenOut),
+      swapper: utils.getAddress(body.swapper),
+      tokenIn: utils.getAddress(body.tokenIn),
+      tokenOut: utils.getAddress(body.tokenOut),
       amount: BigNumber.from(body.amount),
       type: TradeType[body.type as keyof typeof TradeType],
       numOutputs: body.numOutputs,
+      cosigner: utils.getAddress(body.cosigner),
     });
   }
 
-  constructor(private data: QuoteRequestData) {}
+  constructor(private data: V2QuoteRequestData) {}
 
-  public toJSON(): QuoteRequestDataJSON {
+  public toJSON(): V2QuoteRequestJSON {
     return {
       tokenInChainId: this.tokenInChainId,
       tokenOutChainId: this.tokenOutChainId,
       requestId: this.requestId,
-      swapper: getAddress(this.swapper),
-      tokenIn: getAddress(this.tokenIn),
-      tokenOut: getAddress(this.tokenOut),
+      swapper: utils.getAddress(this.swapper),
+      tokenIn: utils.getAddress(this.tokenIn),
+      tokenOut: utils.getAddress(this.tokenOut),
       amount: this.amount.toString(),
       type: TradeType[this.type],
       numOutputs: this.numOutputs,
+      cosigner: utils.getAddress(this.cosigner),
       ...(this.quoteId && { quoteId: this.quoteId }),
     };
   }
 
-  public toCleanJSON(): QuoteRequestDataJSON {
+  public toCleanJSON(): V2QuoteRequestJSON {
     return {
       tokenInChainId: this.tokenInChainId,
       tokenOutChainId: this.tokenOutChainId,
       requestId: this.requestId,
-      tokenIn: getAddress(this.tokenIn),
-      tokenOut: getAddress(this.tokenOut),
+      tokenIn: utils.getAddress(this.tokenIn),
+      tokenOut: utils.getAddress(this.tokenOut),
       amount: this.amount.toString(),
-      swapper: ethers.constants.AddressZero,
+      swapper: constants.AddressZero,
       type: TradeType[this.type],
       numOutputs: this.numOutputs,
+      cosigner: utils.getAddress(this.cosigner),
       ...(this.quoteId && { quoteId: this.quoteId }),
     };
   }
 
   // return an opposing quote request,
   // i.e. quoting the other side of the trade
-  public toOpposingCleanJSON(): QuoteRequestDataJSON {
+  public toOpposingCleanJSON(): V2QuoteRequestJSON {
     const type = this.type === TradeType.EXACT_INPUT ? TradeType.EXACT_OUTPUT : TradeType.EXACT_INPUT;
     return {
       tokenInChainId: this.tokenOutChainId,
       tokenOutChainId: this.tokenInChainId,
       requestId: this.requestId,
       // switch tokenIn/tokenOut
-      tokenIn: getAddress(this.tokenOut),
-      tokenOut: getAddress(this.tokenIn),
+      tokenIn: utils.getAddress(this.tokenOut),
+      tokenOut: utils.getAddress(this.tokenIn),
       amount: this.amount.toString(),
-      swapper: ethers.constants.AddressZero,
+      swapper: constants.AddressZero,
       // switch tradeType
       type: TradeType[type],
       numOutputs: this.numOutputs,
+      cosigner: utils.getAddress(this.cosigner),
       ...(this.quoteId && { quoteId: this.quoteId }),
     };
   }
 
-  public toOpposingRequest(): QuoteRequest {
+  public toOpposingRequest(): V2QuoteRequest {
     const opposingJSON = this.toOpposingCleanJSON();
-    return new QuoteRequest({
+    return new V2QuoteRequest({
       ...opposingJSON,
       amount: BigNumber.from(opposingJSON.amount),
       type: TradeType[opposingJSON.type as keyof typeof TradeType],
@@ -136,6 +126,10 @@ export class QuoteRequest {
 
   public get numOutputs(): number {
     return this.data.numOutputs;
+  }
+
+  public get cosigner(): string {
+    return this.data.cosigner;
   }
 
   public get quoteId(): string | undefined {
