+++ lib/handlers/quote-v2/hard/schema.ts
@@ -1,0 +1,22 @@
+import Joi from 'joi';
+
+import { FieldValidator } from '../../../util/validator';
+
+/* Hard quote request from user */
+export const V2HardQuoteRequestBodyJoi = Joi.object({
+  requestId: FieldValidator.requestId.required(),
+  quoteId: FieldValidator.uuid.optional(),
+  encodedInnerOrder: Joi.string().required(),
+  innerSig: FieldValidator.rawSignature.required(),
+  tokenInChainId: FieldValidator.chainId.required(),
+  tokenOutChainId: FieldValidator.chainId.required(),
+});
+
+export type V2HardQuoteRequestBody = {
+  requestId: string;
+  quoteId?: string;
+  encodedInnerOrder: string;
+  innerSig: string;
+  tokenInChainId: number;
+  tokenOutChainId: number;
+};
