+++ lib/handlers/quote-v2/index.ts
@@ -1,3 +1,3 @@
-export * from './handler';
+export * from './indicative/handler';
 export * from './injector';
 export * from './schema';
